requirements.txt
nextcord
asyncio
python-dotenv
pynacl
requests
snscrape